zbuffer

+++ Box.cs
@@ -3,10 +3,10 @@ namespace tiny_renderer
 {
     public struct Box
     {
-        public float minX;
-        public float minY;
-        public float maxX;
-        public float maxY;
+        public int minX;
+        public int minY;
+        public int maxX;
+        public int maxY;
     }
 }
